requirements.txt
requests
flask
flask-session
werkzeug
pymongo
flask-socketio
google-auth
google-auth-oauthlib
opencage
python-dotenv
